Backend/server_backend_code/requirements.txt
transformers
torch>=2.0.0
accelerate
safetensors
pillow
torchvision
tqdm
tensorizer>=2.7.0
diffusers
pydantic
matplotlib
numpy
